+++ js/ui/drawer/ui.drawer.rendering.strategy.js
